Correctly generate a temporary file name for bed files
This fixes a problem where sep_chrom_bed could easily create clashed file
names, while working to actually not create clashed file names. The problem
is that, rather than using tempfile, the code was using seconds since the
Unix epoch and then, if it looked like a clash was possible, adding a random
number to the end of it.

The problem here is that, mixed with a fork (which it would do), it would
create multiple forks that all followed the sane random sequence. See:

https://metacpan.org/pod/Parallel::ForkManager#USING-RAND()-IN-FORKED-PROCESSES

for why. Long story short: the code that would try and ensure there wasn't a
filename clash would almost guarantee that there was.

So this change switches to using an actual temp file name generation
function to create the temporary bed file names.